+++ packages/web/src/fixtures/dev-kit/hooks.ts
@@ -1,21 +1,10 @@
-import { newClient } from './client'
+import { newClient, getMyStakingAmount, getRewardsAmount, getTotalStakingAmount } from './client'
 import { SWRCachePath } from './cache-path'
 import { addresses } from '@devprtcl/dev-kit-js'
 import { UnwrapFunc, toNaturalNumber } from 'src/fixtures/utility'
-import { getAccountAddress } from 'src/fixtures/wallet/utility'
 import useSWR from 'swr'
 import { message } from 'antd'
 import { useState } from 'react'
-
-const getRewardsAmount = async (propertyAddress: string) => {
-  const client = newClient()
-  if (client) {
-    return client
-      .withdraw(await client.registry(addresses.eth.main.registry).withdraw())
-      .getRewardsAmount(propertyAddress)
-  }
-  return undefined
-}
 
 export const useGetTotalRewardsAmount = (propertyAddress: string) => {
   const { data, error } = useSWR<UnwrapFunc<typeof getRewardsAmount>, Error>(
@@ -48,14 +37,6 @@ export const useWithdrawHolderReward = () => {
   return { withdraw, isLoading, error }
 }
 
-const getTotalStakingAmount = async (proepertyAddress: string) => {
-  const client = newClient()
-  if (client) {
-    return client.lockup(await client.registry(addresses.eth.main.registry).lockup()).getPropertyValue(proepertyAddress)
-  }
-  return undefined
-}
-
 export const useGetTotalStakingAmount = (propertyAddress: string) => {
   const { data, error } = useSWR<UnwrapFunc<typeof getTotalStakingAmount>, Error>(
     SWRCachePath.getTotalStakingAmount(propertyAddress),
@@ -63,17 +44,6 @@ export const useGetTotalStakingAmount = (propertyAddress: string) => {
     { onError: err => message.error(err.message) }
   )
   return { totalStakingAmount: data ? toNaturalNumber(data) : undefined, error }
-}
-
-const getMyStakingAmount = async (propertyAddress: string) => {
-  const client = newClient()
-  const accountAddress = getAccountAddress()
-  if (client && accountAddress) {
-    return client
-      .lockup(await client.registry(addresses.eth.main.registry).lockup())
-      .getValue(propertyAddress, accountAddress)
-  }
-  return undefined
 }
 
 export const useGetMyStakingAmount = (propertyAddress: string) => {
